requirements.txt
deepagents
tavily-python
multi-search-api
python-dotenv
rich
pyyaml
fastapi
uvicorn[standard]
jinja2